requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
lightgbm>=3.3.0
transformers>=4.30.0
torch>=2.0.0
pytz>=2023.3
scipy>=1.10.0
matplotlib>=3.7.0
pyyaml>=6.0
pyarrow>=12.0.0
